# Query frontend optimisations.