We were seeing the query frontend adding 500ms - 1s to some queries, for instance when the query results JSON is large (1-5MB).  The query itself is only taking ~200ms on the querier.

- Add various spans to the query frontend to see where the latency was.  Its mostly in the JSON marshaling / unmarshaling.
- Switch some more JSON code over to jsoniter, and write a zero-copy marshal / unmarshal function for Samples.

End result is reducing query latency of this particular case by about 50%.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>