@tomwilkie 2019-01-02
Before:
```
$ JSON=$(pwd)/results.json go test -bench BenchmarkMarshalling -run ^$ ./pkg/querier/frontend/
goos: darwin
goarch: amd64
pkg: github.com/cortexproject/cortex/pkg/querier/frontend
BenchmarkMarshalling-8              5     301852363 ns/op
PASS
ok      github.com/cortexproject/cortex/pkg/querier/frontend    2.769s
```

After:
```
$ JSON=$(pwd)/results.json go test -bench BenchmarkMarshalling -run ^$ ./pkg/querier/frontend/ 
goos: darwin
goarch: amd64
pkg: github.com/cortexproject/cortex/pkg/querier/frontend
BenchmarkMarshalling-8             20      67739227 ns/op
PASS
ok      github.com/cortexproject/cortex/pkg/querier/frontend    1.465s
```